Add a network policy to allow backstage-sonata traffic
Set a network policy that will allow ingress traffic only from RHDH
backstage to any sonata workflow, and any sonataflow services ,
data-index and job service.

All pods in namespace where this policy is applied which are NOT
selected by the pod selector, i.e not a sonataflow runtime and not a
sonataflow service, would be accessible, unless other rules are defined
for them.

Signed-off-by: Roy Golan <rgolan@redhat.com>